refactor: Bookmark 리팩토링

+++ src/main/java/com/ssafy/nagne/service/BookmarkService.java
@@ -1,9 +1,13 @@
 package com.ssafy.nagne.service;
 
-import static lombok.Lombok.checkNotNull;
+import static com.google.common.base.Preconditions.checkNotNull;
 
+import com.ssafy.nagne.error.DuplicateException;
+import com.ssafy.nagne.error.NotFoundException;
 import com.ssafy.nagne.repository.BookmarkRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -14,11 +18,11 @@ public class BookmarkService {
 
     private final BookmarkRepository bookmarkRepository;
 
-    public boolean save(Long userId, Long articleId) {
+    public boolean bookmark(Long userId, Long articleId) {
         checkNotNull(userId, "userId must be provided");
         checkNotNull(articleId, "articleId must be provided");
 
-        return bookmarkRepository.save(userId, articleId) == 1;
+        return save(userId, articleId);
     }
 
     @Transactional(readOnly = true)
@@ -29,9 +33,27 @@ public class BookmarkService {
         return bookmarkRepository.check(userId, articleId);
     }
 
-    public boolean delete(Long userId, Long articleId) {
+    public boolean cancel(Long userId, Long articleId) {
         checkNotNull(userId, "userId must be provided");
         checkNotNull(articleId, "articleId must be provided");
+
+        return delete(userId, articleId);
+    }
+
+    private boolean save(Long userId, Long articleId) {
+        try {
+            return bookmarkRepository.save(userId, articleId) == 1;
+        } catch (DuplicateKeyException e) {
+            throw new DuplicateException("already bookmarked");
+        } catch (DataIntegrityViolationException e) {
+            throw new NotFoundException("Could not found article for " + articleId);
+        }
+    }
+
+    private boolean delete(Long userId, Long articleId) {
+        if (!check(userId, articleId)) {
+            throw new DuplicateException("already canceled");
+        }
 
         return bookmarkRepository.delete(userId, articleId) == 1;
     }
